+++ hannah_webui/extensions.py
@@ -46,6 +46,7 @@ TRUST_LEVELS = {
     "edit_user": 10,
     "delete_user": 10,
     "link_resident": 10,
+    "edit_presence_sources": 10,
     "list_activity_log": 0,
     "filter_activity_log": 10,
     "list_messages": 0,